int/uint -> isize/usize

+++ src/cut/cut.rs
@@ -53,7 +53,7 @@ fn list_to_ranges(list: &str, complement: bool) -> Result<Vec<Range>, String> {
 
 fn cut_bytes<R: Reader>(reader: R,
                         ranges: &Vec<Range>,
-                        opts: &Options) -> int {
+                        opts: &Options) -> isize {
     use buffer::Bytes::Select;
     use buffer::Bytes::Selected::{NewlineFound, Complete, Partial, EndOfFile};
 
@@ -134,7 +134,7 @@ fn cut_bytes<R: Reader>(reader: R,
 
 fn cut_characters<R: Reader>(reader: R,
                              ranges: &Vec<Range>,
-                             opts: &Options) -> int {
+                             opts: &Options) -> isize {
     let mut buf_in = BufferedReader::new(reader);
     let mut out = BufferedWriter::new(stdio::stdout_raw());
 
@@ -195,7 +195,7 @@ fn cut_characters<R: Reader>(reader: R,
 struct Searcher<'a> {
     haystack: &'a [u8],
     needle: &'a [u8],
-    position: uint
+    position: usize 
 }
 
 impl<'a> Searcher<'a> {
@@ -209,9 +209,9 @@ impl<'a> Searcher<'a> {
 }
 
 impl<'a> Iterator for Searcher<'a> {
-    type Item = (uint, uint);
+    type Item = (usize, usize);
 
-    fn next(&mut self) -> Option<(uint, uint)> {
+    fn next(&mut self) -> Option<(usize, usize)> {
         if self.needle.len() == 1 {
             for offset in range(self.position, self.haystack.len()) {
                 if self.haystack[offset] == self.needle[0] {
@@ -242,7 +242,7 @@ fn cut_fields_delimiter<R: Reader>(reader: R,
                                    ranges: &Vec<Range>,
                                    delim: &String,
                                    only_delimited: bool,
-                                   out_delim: &String) -> int {
+                                   out_delim: &String) -> isize {
     let mut buf_in = BufferedReader::new(reader);
     let mut out = BufferedWriter::new(stdio::stdout_raw());
 
@@ -316,7 +316,7 @@ fn cut_fields_delimiter<R: Reader>(reader: R,
 
 fn cut_fields<R: Reader>(reader: R,
                          ranges: &Vec<Range>,
-                         opts: &FieldOptions) -> int {
+                         opts: &FieldOptions) -> isize {
     match opts.out_delimeter {
         Some(ref delim) => {
             return cut_fields_delimiter(reader, ranges, &opts.delimiter,
@@ -395,7 +395,7 @@ fn cut_fields<R: Reader>(reader: R,
     0
 }
 
-fn cut_files(mut filenames: Vec<String>, mode: Mode) -> int {
+fn cut_files(mut filenames: Vec<String>, mode: Mode) -> isize {
     let mut stdin_read = false;
     let mut exit_code = 0;
