+++ src/fmt/parasplit.rs
@@ -15,8 +15,8 @@ use FileOrStdReader;
 use FmtOptions;
 
 #[inline(always)]
-fn char_width(c: char) -> uint {
-    if (c as uint) < 0xA0 {
+fn char_width(c: char) -> usize {
+    if (c as usize) < 0xA0 {
         // if it is ASCII, call it exactly 1 wide (including control chars)
         // calling control chars' widths 1 is consistent with OpenBSD fmt
         1
@@ -59,10 +59,10 @@ impl Line {
 #[derive(Show)]
 struct FileLine {
     line       : String,
-    indent_end : uint,     // the end of the indent, always the start of the text
-    pfxind_end : uint,     // the end of the PREFIX's indent, that is, the spaces before the prefix
-    indent_len : uint,     // display length of indent taking into account tabs
-    prefix_len : uint,     // PREFIX indent length taking into account tabs
+    indent_end : usize,     // the end of the indent, always the start of the text
+    pfxind_end : usize,     // the end of the PREFIX's indent, that is, the spaces before the prefix
+    indent_len : usize,     // display length of indent taking into account tabs
+    prefix_len : usize,     // PREFIX indent length taking into account tabs
 }
 
 // iterator that produces a stream of Lines from a file
@@ -77,7 +77,7 @@ impl<'a> FileLines<'a> {
     }
 
     // returns true if this line should be formatted
-    fn match_prefix(&self, line: &str) -> (bool, uint) {
+    fn match_prefix(&self, line: &str) -> (bool, usize) {
         if !self.opts.use_prefix { return (true, 0u); }
 
         FileLines::match_prefix_generic(self.opts.prefix.as_slice(), line, self.opts.xprefix)
@@ -93,7 +93,7 @@ impl<'a> FileLines<'a> {
         }
     }
 
-    fn match_prefix_generic(pfx: &str, line: &str, exact: bool) -> (bool, uint) {
+    fn match_prefix_generic(pfx: &str, line: &str, exact: bool) -> (bool, usize) {
         if line.starts_with(pfx) {
             return (true, 0);
         }
@@ -112,7 +112,7 @@ impl<'a> FileLines<'a> {
         (false, 0)
     }
 
-    fn compute_indent(&self, string: &str, prefix_end: uint) -> (uint, uint, uint) {
+    fn compute_indent(&self, string: &str, prefix_end: usize) -> (usize, usize, usize) {
         let mut prefix_len = 0;
         let mut indent_len = 0;
         let mut indent_end = 0;
@@ -199,11 +199,11 @@ impl<'a> Iterator<Line> for FileLines<'a> {
 pub struct Paragraph {
         lines       : Vec<String>,  // the lines of the file
     pub init_str    : String,       // string representing the init, that is, the first line's indent
-    pub init_len    : uint,         // printable length of the init string considering TABWIDTH
-        init_end    : uint,         // byte location of end of init in first line String
+    pub init_len    : usize,        // printable length of the init string considering TABWIDTH
+        init_end    : usize,        // byte location of end of init in first line String
     pub indent_str  : String,       // string representing indent
-    pub indent_len  : uint,         // length of above
-        indent_end  : uint,         // byte location of end of indent (in crown and tagged mode, only applies to 2nd line and onward)
+    pub indent_len  : usize,        // length of above
+        indent_end  : usize,        // byte location of end of indent (in crown and tagged mode, only applies to 2nd line and onward)
     pub mail_header : bool          // we need to know if this is a mail header because we do word splitting differently in that case
 }
 
@@ -243,7 +243,7 @@ impl<'a> ParagraphStream<'a> {
                 // header field must be nonzero length
                 if colon_posn == 0 { return false; }
 
-                return l_slice.slice_to(colon_posn).chars().all(|x| match x as uint {
+                return l_slice.slice_to(colon_posn).chars().all(|x| match x as usize {
                     y if y < 33 || y > 126 => false,
                     _ => true
                 });
@@ -446,13 +446,13 @@ impl<'a> ParaWords<'a> {
 struct WordSplit<'a> {
     opts       : &'a FmtOptions,
     string     : &'a str,
-    length     : uint,
-    position   : uint,
+    length     : usize,
+    position   : usize,
     prev_punct : bool
 }
 
 impl<'a> WordSplit<'a> {
-    fn analyze_tabs(&self, string: &str) -> (Option<uint>, uint, Option<uint>) {
+    fn analyze_tabs(&self, string: &str) -> (Option<usize>, usize, Option<usize>) {
         // given a string, determine (length before tab) and (printed length after first tab)
         // if there are no tabs, beforetab = -1 and aftertab is the printed length
         let mut beforetab = None;
@@ -494,10 +494,10 @@ impl<'a> WordSplit<'a> {
 
 pub struct WordInfo<'a> {
     pub word           : &'a str,
-    pub word_start     : uint,
-    pub word_nchars    : uint,
-    pub before_tab     : Option<uint>,
-    pub after_tab      : uint,
+    pub word_start     : usize,
+    pub word_nchars    : usize,
+    pub before_tab     : Option<usize>,
+    pub after_tab      : usize,
     pub sentence_start : bool,
     pub ends_punct     : bool,
     pub new_line       : bool
