+++ src/sync/sync.rs
@@ -28,7 +28,7 @@ mod platform {
         fn sync() -> libc::c_void;
     }
 
-    pub unsafe fn do_sync() -> int {
+    pub unsafe fn do_sync() -> isize {
         sync();
         0
     }
@@ -177,7 +177,7 @@ fn help(program: &str, options: &[getopts::OptGroup]) {
     print!("{}", usage("Force changed blocks to disk, update the super block.", options));
 }
 
-fn uusync() -> int {
+fn uusync() -> isize {
     unsafe {
         platform::do_sync()
     }
